requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pymysql
alembic
python-multipart
httpx


